Xic:  fix compile issues (int64_t vs intptr_t) on OSX.

+++ xic/include/fio_cxfact.h
@@ -129,7 +129,7 @@ struct bytefact_t
             return (b_records);
         }
 
-    uintptr_t bytes_inuse()     { return (b_inuse); }
+    uint64_t bytes_inuse()      { return (b_inuse); }
     unsigned int allocations()  { return (b_allocations); }
 
 protected:
@@ -138,7 +138,7 @@ protected:
     unsigned int b_allocated;   // size of array
     unsigned int b_blkused;     // current offset into block
     unsigned int b_allocations; // number of allocations
-    uintptr_t b_inuse;          // bytes in use
+    uint64_t b_inuse;           // bytes in use
     unsigned int b_records;     // user's record count
     bool b_full;                // out of tickets
 };
@@ -234,7 +234,7 @@ struct cmp_bytefact_t
             return (b_records);
         }
 
-    uintptr_t bytes_inuse()     { return (b_inuse); }
+    uint64_t bytes_inuse()      { return (b_inuse); }
     unsigned int allocations()  { return (b_allocations); }
 
     static void set_cmp_threshold(unsigned int t) { b_cmp_threshold = t; }
@@ -287,7 +287,7 @@ private:
     unsigned int b_allocated;   // size of array
     unsigned int b_blkused;     // current offset into block
     unsigned int b_allocations; // number of allocations
-    uintptr_t b_inuse;          // bytes in use
+    uint64_t b_inuse;           // bytes in use
     unsigned int b_records;     // user's record count
     bool b_full;                // out of tickets
 
@@ -662,7 +662,7 @@ struct cxfact_t
     bool save_space(ticket_t) { }
 #endif
 
-    void cref_count(unsigned int *ncrefs, uintptr_t *bytes_inuse)
+    void cref_count(unsigned int *ncrefs, uint64_t *bytes_inuse)
         {
             *ncrefs = crstab->add_count(0);
             *bytes_inuse = crstab->bytes_inuse();
